add file exit app
