+++ src/utils/mod.rs
@@ -3,3 +3,4 @@ pub mod pop_up_message;
 pub mod logs;
 pub mod changelog;
 pub mod encrypt_token;
+pub mod exit_app;
